% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet_exp.R
\name{glmnet_exp}
\alias{glmnet_exp}
\title{fit glmnet model for exponetially distributed response data}
\usage{
glmnet_exp(A, b, alpha = 1, num_lambda = 100L, glm_type = 1L,
  max_iter = 100L, abs_tol = 1e-04, rel_tol = 0.01,
  normalize_grad = FALSE, k_fold = 5L)
}
\arguments{
\item{A}{The matrix of independent variables}

\item{b}{The vector of response variables}

\item{alpha}{the coefficient of elastic net regularizer (1 means lasso)}

\item{num_lambda}{size of the lambda grid}

\item{glm_type}{type of glm model, 1 is exponential, 2 is gamma (not implemented yet)}

\item{max_iter}{max number of iteration for the prox grad descent optimizer}

\item{abs_tol}{absolute error threshold for the pgd optimizer}

\item{rel_tol}{relative error threshold for the pgd optimizer (not used for vanilla PGD)}

\item{normalize_grad}{swtich for whether to normalize the gradient or not}

\item{k_fold}{the number of folds for cross validation}
}
\value{
vector of optimal coefficient for the glm model
}
\description{
fit glmnet model for exponetially distributed response data
}
